fix: move Shopify shop parameter persistence logic to a separate plugin file
